% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_layout.R
\name{layout_pk_attraction}
\alias{layout_pk_attraction}
\title{Layout attraction with Hook's law}
\usage{
layout_pk_attraction(states, edges, k_hook = 1, dist_spring = 1)
}
\arguments{
\item{states}{A data.frame with the states in the model.  The data.frame must
have a column called \code{names}, \code{x}, and \code{y} with the state
names and positions.}

\item{k_hook}{Spring constant for Hook's law}

\item{dist_spring}{Zero deformation distance for the spring}
}
\description{
Layout attraction with Hook's law
}
