% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_layout.R
\name{layout_pk_repulsion}
\alias{layout_pk_repulsion}
\title{Layout repulsion with electrostatic repulsion equation (Coulomb's law)}
\usage{
layout_pk_repulsion(states, k_electro = 1)
}
\arguments{
\item{states}{A data.frame with the states in the model.  The data.frame must
have a column called \code{names}, \code{x}, and \code{y} with the state
names and positions.}

\item{k_electro}{Coulomb constant}
}
\description{
Layout repulsion with electrostatic repulsion equation (Coulomb's law)
}
