% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_network.R
\name{pk_model_network}
\alias{pk_model_network}
\alias{as_pk_model_network}
\title{Generate a pk_model_network object for subsequent layout and plotting}
\usage{
pk_model_network(
  state,
  state_type = NA_character_,
  from,
  to,
  edge_type = "mass"
)

as_pk_model_network(states, edges)
}
\arguments{
\item{state}{The model state names as a character vector}

\item{state_type}{The type of state in the model (currently unused)}

\item{edge_type}{The type of edge in the model. Typically, \code{edge_type}
will be \code{"mass"} for mass transfer and \code{"effect"} for an effect
from one compartment to another without a mass transfer}

\item{states}{The model states (typically the compartments in the model)}

\item{edges}{The connectivity of the model}
}
\description{
Generate a pk_model_network object for subsequent layout and plotting
}
\section{Functions}{
\itemize{
\item \code{as_pk_model_network()}: Generate a pk_model_network object from two
data.frames

}}
\seealso{
Other Network setup: 
\code{\link{pk_model_network_edges}()},
\code{\link{pk_model_network_states}()}
}
\concept{Network setup}
