% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_network.R
\name{pk_model_network_edges}
\alias{pk_model_network_edges}
\title{Define all of the states in the model}
\usage{
pk_model_network_edges(from, to, edge_type)
}
\arguments{
\item{edge_type}{The type of edge in the model. Typically, \code{edge_type}
will be \code{"mass"} for mass transfer and \code{"effect"} for an effect
from one compartment to another without a mass transfer}

\item{state}{The model state names as a character vector}
}
\value{
A data.frame with the network edges as \code{from} and \code{to} and
the \code{type} of edge as three columns
}
\description{
Define all of the states in the model
}
\examples{
# For an oral, one-compartment PK model
model_network_edges(
  from=c("depot", "central"),
  to = c("central", "elimination")
)
}
\seealso{
Other Network setup: 
\code{\link{pk_model_network_states}()},
\code{\link{pk_model_network}()}
}
\concept{Network setup}
