% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_network.R
\name{pk_model_network_states}
\alias{pk_model_network_states}
\title{Define all of the states in the model}
\usage{
pk_model_network_states(state, state_type)
}
\arguments{
\item{state}{The model state names as a character vector}

\item{state_type}{The type of state in the model (currently unused)}
}
\value{
A data.frame with the state \code{name} and the \code{type} as two
columns
}
\description{
Define all of the states in the model
}
\examples{
model_network_states(c("depot", "central", "elimination"), type = "pk")
}
\seealso{
Other Network setup: 
\code{\link{pk_model_network_edges}()},
\code{\link{pk_model_network}()}
}
\concept{Network setup}
